requirements.txt
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
black==21.5b2
click==8.0.1
colorama==0.4.4
Django==3.2.3
django-js-asset==1.2.2
django-mptt==0.13.2
djangorestframework==3.12.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
Pillow==8.3.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
regex==2021.4.4
selenium==3.141.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
